+++ R/tm_a_mmrm.R
@@ -167,7 +167,7 @@ template_mmrm_tables <- function(parentname,
                                  paramcd,
                                  show_relative = c("increase", "reduction", "none"),
                                  table_type = "t_mmrm_cov",
-                                 total_label = "All Patients",
+                                 total_label = default_total_label(),
                                  basic_table_args = teal.widgets::basic_table_args()) {
   y <- list()
   ref_arm_val <- paste(ref_arm, collapse = "/")
@@ -522,7 +522,7 @@ tm_a_mmrm <- function(label,
                       conf_level = teal.transform::choices_selected(c(0.95, 0.9, 0.8), 0.95, keep_order = TRUE),
                       plot_height = c(700L, 200L, 2000L),
                       plot_width = NULL,
-                      total_label = "All Patients",
+                      total_label = default_total_label(),
                       pre_output = NULL,
                       post_output = NULL,
                       basic_table_args = teal.widgets::basic_table_args(),
@@ -824,7 +824,8 @@ srv_mmrm <- function(id,
                      ggplot2_args) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   with_filter <- !missing(filter_panel_api) && inherits(filter_panel_api, "FilterPanelAPI")
-  checkmate::assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   shiny::moduleServer(id, function(input, output, session) {
     # Reactive responsible for sending a disable/enable signal
@@ -853,7 +854,7 @@ srv_mmrm <- function(id,
       input,
       output,
       id_arm_var = extract_input("arm_var", parentname), # From UI.
-      data = data[[parentname]],
+      data = shiny::reactive(data()[[parentname]]),
       arm_ref_comp = arm_ref_comp,
       module = "tm_mmrm"
     )
@@ -920,21 +921,19 @@ srv_mmrm <- function(id,
     anl_inputs <- teal.transform::merge_expression_srv(
       datasets = data,
       selector_list = selector_list_without_cov,
-      join_keys = teal.data::join_keys(data),
       merge_function = "dplyr::inner_join"
     )
 
     adsl_merge_inputs <- teal.transform::merge_expression_module(
       datasets = data,
       data_extract = list(arm_var = arm_var),
-      join_keys = teal.data::join_keys(data),
       anl_name = "ANL_ADSL"
     )
 
     anl_q <- shiny::reactive({
-      qenv <- teal.code::new_qenv(tdata2env(data), code = get_code_tdata(data))
-      qenv2 <- teal.code::eval_code(qenv, as.expression(anl_inputs()$expr))
-      teal.code::eval_code(qenv2, as.expression(adsl_merge_inputs()$expr))
+      data() %>%
+        teal.code::eval_code(code = as.expression(anl_inputs()$expr)) %>%
+        teal.code::eval_code(code = as.expression(adsl_merge_inputs()$expr))
     })
 
     # Initially hide the output title because there is no output yet.
@@ -1457,7 +1456,7 @@ srv_mmrm <- function(id,
           card$append_text("Comment", "header3")
           card$append_text(comment)
         }
-        card$append_src(paste(teal.code::get_code(all_q()), collapse = "\n"))
+        card$append_src(teal.code::get_code(all_q()))
         card
       }
       teal.reporter::simple_reporter_srv("simple_reporter", reporter = reporter, card_fun = card_fun)