+++ R/tm_g_ci.R
@@ -376,7 +376,8 @@ srv_g_ci <- function(id, # nolint
                      ggplot2_args) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   with_filter <- !missing(filter_panel_api) && inherits(filter_panel_api, "FilterPanelAPI")
-  checkmate::assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   shiny::moduleServer(id, function(input, output, session) {
     selector_list <- teal.transform::data_extract_multiple_srv(
@@ -407,12 +408,10 @@ srv_g_ci <- function(id, # nolint
       selector_list = selector_list
     )
 
-    anl_q <- shiny::reactive(
-      teal.code::eval_code(
-        object = teal.code::new_qenv(tdata2env(data), code = get_code_tdata(data)),
-        code = as.expression(anl_inputs()$expr)
-      )
-    )
+    anl_q <- shiny::reactive({
+      data() %>%
+        teal.code::eval_code(as.expression(anl_inputs()$expr))
+    })
 
     all_q <- shiny::reactive({
       teal::validate_inputs(iv_r())
@@ -429,10 +428,10 @@ srv_g_ci <- function(id, # nolint
         )
       )
 
-      x_label <- column_annotation_label(data[[attr(x, "dataname")]](), x)
-      y_label <- column_annotation_label(data[[attr(y, "dataname")]](), y)
+      x_label <- column_annotation_label(data()[[attr(x, "dataname")]], x)
+      y_label <- column_annotation_label(data()[[attr(y, "dataname")]], y)
       color_label <- if (length(color)) {
-        column_annotation_label(data[[attr(color, "dataname")]](), color)
+        column_annotation_label(data()[[attr(color, "dataname")]], color)
       } else {
         NULL
       }
@@ -501,7 +500,7 @@ srv_g_ci <- function(id, # nolint
           card$append_text("Comment", "header3")
           card$append_text(comment)
         }
-        card$append_src(paste(teal.code::get_code(all_q()), collapse = "\n"))
+        card$append_src(teal.code::get_code(all_q()))
         card
       }
       teal.reporter::simple_reporter_srv("simple_reporter", reporter = reporter, card_fun = card_fun)