+++ R/tm_g_lineplot.R
@@ -285,7 +285,7 @@ tm_g_lineplot <- function(label,
                           interval = "mean_ci",
                           mid = "mean",
                           whiskers = c("mean_ci_lwr", "mean_ci_upr"),
-                          table = NULL,
+                          table = c("n", "mean_sd", "median", "range"),
                           mid_type = "pl",
                           mid_point_size = c(2, 1, 5),
                           table_font_size = c(4, 2, 6),
@@ -488,7 +488,7 @@ ui_g_lineplot <- function(id, ...) {
               "25% and 75%-ile" = "quantiles",
               "Min - Max" = "range"
             ),
-            selected = c("n", "mean_sd", "median", "range"),
+            selected = a$table,
           )
         )
       )
@@ -524,7 +524,8 @@ srv_g_lineplot <- function(id,
                            ggplot2_args) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   with_filter <- !missing(filter_panel_api) && inherits(filter_panel_api, "FilterPanelAPI")
-  checkmate::assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   shiny::moduleServer(id, function(input, output, session) {
     selector_list <- teal.transform::data_extract_multiple_srv(
@@ -555,13 +556,12 @@ srv_g_lineplot <- function(id,
 
     anl_inputs <- teal.transform::merge_expression_srv(
       datasets = data,
-      join_keys = teal.data::join_keys(data),
       selector_list = selector_list,
       merge_function = "dplyr::inner_join"
     )
 
     anl_q <- shiny::reactive({
-      teal.code::new_qenv(tdata2env(data), code = get_code_tdata(data)) %>%
+      data() %>%
         teal.code::eval_code(as.expression(anl_inputs()$expr))
     })
 
@@ -673,7 +673,7 @@ srv_g_lineplot <- function(id,
           card$append_text("Comment", "header3")
           card$append_text(comment)
         }
-        card$append_src(paste(teal.code::get_code(all_q()), collapse = "\n"))
+        card$append_src(teal.code::get_code(all_q()))
         card
       }
       teal.reporter::simple_reporter_srv("simple_reporter", reporter = reporter, card_fun = card_fun)