+++ R/tm_t_abnormality_by_worst_grade.R
@@ -20,7 +20,7 @@ template_abnormality_by_worst_grade <- function(parentname, # nolint
                                                 worst_low_flag_var = "WGRLOFL",
                                                 worst_flag_indicator = "Y",
                                                 add_total = FALSE,
-                                                total_label = "All Patients",
+                                                total_label = default_total_label(),
                                                 drop_arm_levels = TRUE,
                                                 basic_table_args = teal.widgets::basic_table_args()) {
   assertthat::assert_that(
@@ -328,7 +328,7 @@ tm_t_abnormality_by_worst_grade <- function(label, # nolint
                                               selected = "Y", fixed = TRUE
                                             ),
                                             add_total = TRUE,
-                                            total_label = "All Patients",
+                                            total_label = default_total_label(),
                                             drop_arm_levels = TRUE,
                                             pre_output = NULL,
                                             post_output = NULL,
@@ -498,7 +498,8 @@ srv_t_abnormality_by_worst_grade <- function(id, # nolint
                                              basic_table_args) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   with_filter <- !missing(filter_panel_api) && inherits(filter_panel_api, "FilterPanelAPI")
-  checkmate::assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   shiny::moduleServer(id, function(input, output, session) {
     selector_list <- teal.transform::data_extract_multiple_srv(
@@ -532,19 +533,17 @@ srv_t_abnormality_by_worst_grade <- function(id, # nolint
     anl_inputs <- teal.transform::merge_expression_srv(
       selector_list = selector_list,
       datasets = data,
-      join_keys = teal.data::join_keys(data),
       merge_function = "dplyr::inner_join"
     )
 
     adsl_inputs <- teal.transform::merge_expression_module(
       datasets = data,
-      join_keys = teal.data::join_keys(data),
       data_extract = list(arm_var = arm_var),
       anl_name = "ANL_ADSL"
     )
 
     anl_q <- shiny::reactive({
-      teal.code::new_qenv(tdata2env(data), code = get_code_tdata(data)) %>%
+      data() %>%
         teal.code::eval_code(as.expression(anl_inputs()$expr)) %>%
         teal.code::eval_code(as.expression(adsl_inputs()$expr))
     })
@@ -681,7 +680,7 @@ srv_t_abnormality_by_worst_grade <- function(id, # nolint
           card$append_text("Comment", "header3")
           card$append_text(comment)
         }
-        card$append_src(paste(teal.code::get_code(all_q()), collapse = "\n"))
+        card$append_src(teal.code::get_code(all_q()))
         card
       }
       teal.reporter::simple_reporter_srv("simple_reporter", reporter = reporter, card_fun = card_fun)