+++ R/tm_t_coxreg.R
@@ -688,7 +688,8 @@ srv_t_coxreg <- function(id,
                          basic_table_args) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   with_filter <- !missing(filter_panel_api) && inherits(filter_panel_api, "FilterPanelAPI")
-  checkmate::assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   shiny::moduleServer(id, function(input, output, session) {
     # Observer to update reference and comparison arm input options.
@@ -697,7 +698,7 @@ srv_t_coxreg <- function(id,
       input,
       output,
       id_arm_var = extract_input("arm_var", parentname),
-      data = data[[parentname]],
+      data = shiny::reactive(data()[[parentname]]),
       arm_ref_comp = arm_ref_comp,
       module = "tm_t_coxreg"
     )
@@ -790,13 +791,12 @@ srv_t_coxreg <- function(id,
 
     anl_inputs <- teal.transform::merge_expression_srv(
       datasets = data,
-      join_keys = teal.data::join_keys(data),
       selector_list = selector_list,
       merge_function = "dplyr::inner_join"
     )
 
     anl_q <- shiny::reactive({
-      teal.code::new_qenv(tdata2env(data), code = get_code_tdata(data)) %>%
+      data() %>%
         teal.code::eval_code(as.expression(anl_inputs()$expr))
     })
 
@@ -1077,7 +1077,7 @@ srv_t_coxreg <- function(id,
           card$append_text("Comment", "header3")
           card$append_text(comment)
         }
-        card$append_src(paste(teal.code::get_code(all_q()), collapse = "\n"))
+        card$append_src(teal.code::get_code(all_q()))
         card
       }
       teal.reporter::simple_reporter_srv("simple_reporter", reporter = reporter, card_fun = card_fun)