+++ R/tm_t_events_summary.R
@@ -40,7 +40,7 @@ template_events_summary <- function(anl_name,
                                     aeseq_var = "AESEQ",
                                     llt = "AEDECOD",
                                     add_total = TRUE,
-                                    total_label = "All Patients",
+                                    total_label = default_total_label(),
                                     count_subj = TRUE,
                                     count_pt = TRUE,
                                     count_events = TRUE) {
@@ -268,7 +268,10 @@ template_events_summary <- function(anl_name,
   if (add_total) {
     layout_anl_list <- add_expr(
       layout_anl_list,
-      quote(rtables::add_overall_col(label = "All Patients"))
+      substitute(
+        expr = rtables::add_overall_col(label = tot_label),
+        env = list(tot_label = total_label)
+      )
     )
   }
 
@@ -597,7 +600,7 @@ tm_t_events_summary <- function(label,
                                   fixed = TRUE
                                 ),
                                 add_total = TRUE,
-                                total_label = "All Patients",
+                                total_label = default_total_label(),
                                 count_subj = TRUE,
                                 count_pt = TRUE,
                                 count_events = TRUE,
@@ -788,7 +791,8 @@ srv_t_events_summary <- function(id,
                                  basic_table_args) {
   with_reporter <- !missing(reporter) && inherits(reporter, "Reporter")
   with_filter <- !missing(filter_panel_api) && inherits(filter_panel_api, "FilterPanelAPI")
-  checkmate::assert_class(data, "tdata")
+  checkmate::assert_class(data, "reactive")
+  checkmate::assert_class(shiny::isolate(data()), "teal_data")
 
   shiny::moduleServer(id, function(input, output, session) {
     data_extract_vars <- list(
@@ -824,19 +828,17 @@ srv_t_events_summary <- function(id,
     anl_inputs <- teal.transform::merge_expression_srv(
       datasets = data,
       selector_list = selector_list,
-      join_keys = teal.data::join_keys(data),
       merge_function = "dplyr::inner_join"
     )
 
     adsl_inputs <- teal.transform::merge_expression_module(
       datasets = data,
       data_extract = Filter(Negate(is.null), list(arm_var = arm_var, dthfl_var = dthfl_var, dcsreas_var = dcsreas_var)),
-      join_keys = teal.data::join_keys(data),
       anl_name = "ANL_ADSL"
     )
 
     anl_q <- shiny::reactive({
-      teal.code::new_qenv(tdata2env(data), code = get_code_tdata(data)) %>%
+      data() %>%
         teal.code::eval_code(as.expression(anl_inputs()$expr)) %>%
         teal.code::eval_code(as.expression(adsl_inputs()$expr))
     })
@@ -982,7 +984,7 @@ srv_t_events_summary <- function(id,
           card$append_text("Comment", "header3")
           card$append_text(comment)
         }
-        card$append_src(paste(teal.code::get_code(table_q()), collapse = "\n"))
+        card$append_src(teal.code::get_code(table_q()))
         card
       }
       teal.reporter::simple_reporter_srv("simple_reporter", reporter = reporter, card_fun = card_fun)